shell: move the Failed tests into yield.test.ts, tidy the reject path
The tests exercise the Yield::run trampoline, so they live with the
other trampoline tests. Adds a pipeline case. The exception-to-error
conversion has one fallback, and finish() asserts that it cannot run
after a rejection.
